Adding site_tests to client directory
The directory site_tests was conceived long ago as a
place to store tests internal to organizations, that
could not become upstream. Now in Fedora installs,
the absence of this directory will make the code to
throw an exception due to its absence. So let's include
it in the source tree, and install it with setup.py.

CC: Martin Krizek <mkrizek@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>