client.shared.base_job: Use autotest_top_path when no AUTODIR is set
During test of the next branch, I've realized the AUTODIR
environment variable is not set on autoserv jobs, hence,
failing any server job. So, when AUTODIR is not set
(server side job), use the global_config.ini value of
autotest_top_path.

This fixes a bug that was preventing server side jobs to
work on next.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>