server.autotest_remote: Transfer and clean state files properly
The autotest_remote code assumes the state of remote
control files is simply path/to/control + ".state". This
causes problem with our patches moving state file creation
to the test output dir set on global_config.ini, with
fall back to the autotest tmpdir area.

Fix the code, creating new attributes for the _BaseRun
class, that will make it easier to control and keep
consistency of the state file locations.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>