web frontend: Point kernel spec docs on job create page
As we've had some confusion recently about how to specify
a kernel properly in the job creation interfaces, let's
put a link to a documentation page on the wiki that explains
what types of strings are valid and what are not. This
replaces the old, terse string in the web interface.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>