client.cmdparser: add fetch command
When playing with my changes for the Git Fetcher,
I thought it would be nice if we could just 'fetch'
the control file remotely.

This would make it easier for folks to create
random tests and host them, while letting others
just grab them with a url pointer (either git: or
http:).

It creates a new area to store the fetched tests
(based on the 'test_out_dir' variable) and hooks
this area into the 'list' and 'run' commands.

Tested by running:

autotest-local fetch git://<internal-git-site> testsuite/control.simple

and then run it with

autotest-local run testsuite/control.simple

(the control file adds the remote repo to fetch the tests from)

In order to completely abstract the use of the
job.add_repo to control files, just set fetch_location
in global_config.ini.

Signed-off-by: Don Zickus <dzickus@redhat.com>