client.cmdparser: Use the logging system
When testing the new autotest-local fetch command, I've
noticed that a *lot* of debug output is thrown, which
might be confusing.

So, make the CmdParser code to use the logging code
used by other autotest entry points, making the INFO
level the default. This reduces the amount of
output of autotest-local using the fetch action
drastically.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>