Fix repair of systems with install_server
When we are running repair on a system, we aren't passed a profile like
we are with autoserv jobs normally in the machine name. So self.profile
for the Host object comes back as ''. With the code as-is, that will
result in install_machine() returning immediately back to the repair
code and verify complaining with:

07/08 06:45:32 INFO |base_class:0407| Trying to reinstall the machine
07/08 06:45:32 WARNI|base_class:0360| Already done this (<bound method [machine name redacted].machine_install of <remote host: [machine name redacted], profile: >>) repair procedure, re-raising the original exception.

I think a simpler choice is that if we still don't have a profile after
self.profile is assigned, let's assume that the install system can
figure it out? And if it is unable to (usually do to a misconfiguration
of Cobbler backend), thrown an exception.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>