virt.installer: Register keyval with installed sw versions
In order to display the appropriate virt software versions
installed in the autotest results email, we can register
versions like the precise git repo definition used to
build that software, or all the packages from Brew/Koji
that were installed.

This is attained by creating an additional method,
write_version_keyval, that gets executed after the
installation was performed. The first Installer class
that gets a get_version method is GitRepoInstaller.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>