client/virt/tests/rv_connect.py: Prevent killing remote_viewer after test finish
There was a bug which caused gnome-session kill after the test ends. This caused that remote-viewer was killed too.
NOHUP had no influence to change this behavior. DISOWN gnome-session and remote-viewer processes works time to time.

This patch removes get_pid() method which was needed for disown gnome-session and remote-viewer process. And adds disown -ar to all processes running on VM.

After that remote-viewer is running and it's possible to call for example rv_disconnect test.

Signed-off-by: Vaclav Ehrlich <vehrlich@redhat.com>
Signed-off-by: Marian Krcmarik <mkrcmari@redhat.com>