client.harness_autoserv: Make AutoservFetcher call parent's init()
The AutoservFetcher object constructor used to not call
its parent constructor. However, the work to support
a GitFetcher moved some methods to common code, and now
it is necessary to call install_pkg_setup, that does
require the setup made by the parent's constructor.

This patch fixes the following problem, when running
a simple sleeptest using the web interface:

07/30 17:31:44 DEBUG|  parallel:0030| AttributeError: 'AutoservFetcher' object has no attribute 'pkgmgr'

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>