client.shared.base_packages: pass full package name
Restore the old code behavior of passing the full
tarball name to the underlying code, and only parse
the tarball name in case we need the package name
and type back, such as we have with git packager.

This resolves a regression using the autoserv packager
as a package provider for autotest jobs (the default
using the web interface).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>