client.shared.test: Copy patch files to parent of test.srcdir
A bunch of autotest tests do rely on using patch files
to fix/tweak an upstream shipped test tarball. With the
new schema 'test module read only', a series of expectations
of the tests (patch file located on parent of test.srcdir)
is being broken. So copy each patch file found on the test
module to the test tmp area, parent of srcdir.

This patch fixes tests such as bonnie, that rely on such
patches for proper test suite execution.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>