xfstests: experiment with software manager and distro specific deps
This is an experiment, quite ad-hoc way of specifying depencies for
this test suite. My intentions are to stir discussion around a
unified way of providing dependency info, and of course, install
them on demand.

Nevertheless, this change should not affect users of distros other
than Fedora (which has a list of dependencies), and will let users
of this distro run this test more smoothly.

Signed-off-by: Cleber Rosa <crosa@redhat.com>