client.tests: Add new autotest_regression module
Add a new autotest_regression module, tailored to
perform some quick, basic checks against the autotest
tree. It leverages the new install_distro_packages
API to try installing the appropriate package in
some popular distros.

The plan is to expand this check routine, to include
things like basic automated assessment of the health
of the web and cli interfaces.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>