Introduce deps.py and make iso9660.py use it
The 'deps' client library was born out of iso9660.py, which contained
a bunch of functions to check for the presence of executables and
the current process hability to execute them.

This has now been separated into a general purpose library that hopes
to concentrate various test dependencies checks. The first inclusion
to the list are packages.

For some types of dependencies, such as packages, we want to allow on
an opt-in basis, their automatic resolution, that is, the automatic
package installation. The same idea could be applied to other deps
that are simple to deal with, such as mounting local or remote
filesystems.

Signed-off-by: Cleber Rosa <crosa@redhat.com>