virt.unattended install: add syslog logging support
This way we capture the installation progress and
have a chronologically sound co-relation between
host and guest events. By default, the feature will
be turned off, as it adds a *lot* of verbosity to
the test logs, and for newer guests such as recent
Fedora, we have better logging through the virtio
console.

Signed-off-by: Cleber Rosa <crosa@redhat.com>