scheduler.scheduler_models: Do not order by reason anymore
When fetching the results of the tests executed during
a job, remove the order by reason, previously used to
group similar failures. This way, the pass reports and
the failure reports will be ordered by execution order,
which is more interesting with the new report.

Also, show the job URL right after the job summary.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>