scheduler.monitor_db_watcher: Figure out results dir and scheduler path
This is another packaging related fix, we can figure out
autotest-scheduler path with os_dep in case of system wide
install, if not there fall back to the original way of
determining both scheduler and results directory.

It needs some testing to see if it doesn't break stuff I
didn't anticipate.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>