Reverting changes to conmux
Nish reported problems in conmux that can be traced
back to commit

f072c190b171a756417d54c698adf1499069cbf0

So to avoid breakages for existing users of the code,
that commit will be reverted. We'll get back to it when
it is possible to find out an environment where we can
get conmux more thoroughly tested.

CC: Philipp Seiler <seiler@b1-systems.de>
CC: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>