database: allow to run unittests against other database backends.
By using the new database management module.

WARNING: it's currently necessary to edit the code to run the
unittests against other database backends. This is done on purpose
because running the database unittests involve creating and dropping
the database and may result in data loss if pointed to your production
database.

Signed-off-by: Cleber Rosa <crosa@redhat.com>