component isolation: introduce models and database schema
The so called component isolation feature needs to keep track of
more information than we usually do with the AFE/TKO models.

I've decided to loose the afe_ and tko_ suffixes on the database
tables because there's a trend towards uniting the database. Also,
the component isolation deals with some aspects and models on both
AFE and TKO, so it doesn't make sense to pick a side.

If you have a running database, those changes should be applied with:

   $ cd AUTOTEST_ROOT/frontend
   $ ./manage.py migrate afe
   $ ./manage.py migrate tko

If you're creating a database from scratch, no extra step is needed.

This also preloads the database with "unknown" entries.

The unknown distro will be used either when the test does not
register or is able to detect the linux distro it's running on.

The unknown software kind and arch are also used as fallbacks
when not properly recorded.

Changes from v1:
   * add usage info on software kind name field
   * new line clean up on auto generated migration file

Signed-off-by: Cleber Rosa <crosa@redhat.com>