client/tools/boottool.py: Work around bug that resets Grubby.path
I'm still chasing a bug that made boottool to fail when
setting up the autotest boot entry on a guest, but what
the bug does is to reset Grubby.path to None, making
calls to grubby to fail since the executable is set to
None as well.

Create a _set_path() method that takes care of setting
the grubby path, and call it in all relevant places,
specially if we're assembling a grubby command and
realize the executable used was reset to None.

CC: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>