frontend: Change the afe import library style
With the recent changes pushed to simplify the autotest
module import 'magic', we noticed that due to some
django idiosyncrasies back then the style of imports
inside the module all refer things like

frontend.afe.*

rather than

autotest.frontend.afe.*

Now that we don't support the top level dir empty on
the import setup anymore, we should refer to the
complete namespace. This fixes a number of db and
frontend related unittests on next.

CC: Eduardo Habkost <ehabkost@redhat.com>
CC: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>