base_syncdata_unittest: Remove bad unittest
This is the actual solution for issue #296.

The unittest with a fake server (that in fact is not
really fake, it's the barrier listen server) was
creating a rogue, paused process at the end of this
unittest run. As it uses base barrier server, it
will cause the barrier unittest to fail the next
time you run the unittests, since this process is
there holding up the barrier ports.

CC: Vinson Lee <vlee@twitter.com>
CC: Jiri Zupka <jzupka@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>