drone_utility: Fix packaging bug
This is a solution for the issue #596.

On a system wide install, the autoserv processes are in
fact called autotest-remote rather than autoserv, and the
scheduler wasn't updating the process set with them.

As a result, in the scheduler refresh loop, the processes
always looked as finished, which makes the scheduler to
try to read its pid file, and as the process is not actually
finished, report the process as it died without writing
exit code. The result is even when a verify is finished
successfuly the verify step is reported as a failure.

The solution is to make the _refresh_processes private
method of the DroneUtility class to accept a list of
processes to be verified, and verify both 'autoserv'
and 'autotest-remote' processes when refreshing the
autoserv processes present.

This should be pushed both to next and the 0.14 branch.

Martin, you may apply the patch directly to the RPM as
an emergency measure.

CC: Martin Krizek <mkrizek@redhat.com>
CC: Paul Whalen <pwhalen@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>