boottool: add build requirements installation for SUSE systems
SUSE systems don't have grubby available from packages, so grubby
must always be built from sources. This should work if the necessary
packages for building grubby are installed.

This patch goes the extra mile and attemps to install those deps
before attempting to build grubby.

This is a fix for issue #597.

CC: Nishanth Aravamudan <nacc@us.ibm.com>
CC: Lucas Meneghel Rodrigues <lmr@redhat.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>