client/job.py: parameterize the sleep time before reboot
Some older machines are taking longer than 5 seconds to sync before
reboot.  To accomodate these machines the sleep time is parameterized.
The default sleep time is 5 seconds, but if additional time is needed,
the sleep_before_reboot variable in global_config.ini can be adjusted
to allow for more time.

Fixed the sleep command so that it is compatible with python 2.4.

Signed-off-by: Cheri Van Pelt <cheri.vanpelt@hp.com>
Tested-by: Julius Gawlas <julius.gawlas@hp.com>