kernelexpand_unittest: override local configuration
If an end-user has configured a local gitweb source or kernel mirrors,
this can lead to the unittest failing, because the tests assume the
settings are the default ones (for comparison to the expected results).
Fix this by overriding the settings values with the default ones.
Without this change, the unittest always fails with local gitweb/mirror
configuration.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>