afe/HostSelector: add "id" to metaObjects
A following change will rely on all objects in the selectedHostData to
have an "id" property to guarantee a way to identify them when rows are
clicked.  Meta-host objects don't have one, so add one that starts
(effectively) at -1 and decrements. This should avoid collisions with
the actual host objects, which all have positive ids.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>