frontend/afe: fix bug with profile selection in create job
There is a bug in the onClick handler in SelectionManager. We assume the
row index is 1:1 with the available table objects. However, this is not
true in two cases, at least:

1) Re-sorting the table causes the "current" row to be sent in, but the
underlying object collection does not change its order.

2) When metahosts are added to the table, they are not present in the
available table.

We therefore end up altering arbitrary available host row objects, which
leads to eventual errors in execution.

To fix this, associated the "id" of the underlying row object with the
widget, and notify listeners onClick that a given "id" has been updated
to a given "profile". Add an implementation of the onClick handler in
HostSelector that updates the selectedHostData, looked-up by that "id".
Stop modifying the SelectionManager's row objects, which represent the
available table's rows.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>