frontend/afe: add support for meta-host objects to have associated profiles
Use an RPC call, just like for specific hosts, for finding all possible
profiles. We let the user shoot themselves in the foot (that is, they
might select a profile that doesn't apply to that server), but that's
the nature of selecting by label -- unless the label is a platform, we
won't be able to do the lookup anyways in the install server.

Changes from v1:
 * Fix up various parameters in the RPC interfaces to be lists rather than
   tuples.
 * Fix-up unittests to pass in lists rather than tuples.
 * Fix logging exception.
 * Fix monitor_db.py call of create_new_job.
 * Fix indentation issue.
 * Refactor code so install server is available for get_num_profiles().

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>