frontend/install_server: remove filter in profile look-up
Depending on your environment, you may not have access to Cobbler to
always manipulate the profile comments. Also, you may not want to ever
limit the user from selecting a profile, and the filtered search is
slower than just grabbing them all. So, for now, simply grab all defined
profiles.

Eventually (working on it now), we should provide an extensible way to
filter the profiles.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>