control files: always reinstall hosts
Locally, I always want to reinstall hosts when I run a job. I imagine
others might also like this. Given that reinstallation depends on
profile availability (of which, "Do_not_install" and "N/A" are valid
depending on the configuration), we can just always reinstall in the
client and server control file templates.

Changes from v1:
 * Fix-up the unit-tests.
 * Adjust machine_install to no throw an exception if no install server is
   configured (just becomes a no-op)

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>