Fixed apache wsgi configuration
Added directive to pass authorization token in wsgi
apache config, without that rpc (including CLI)
was broked and it was not sending the user name
properly since without that HTTP_AUTHORIZATION was
being striped. Also renamed wsgi process group
to be autotest specific to make it easier to
coexist with other apps on the same server.

Signed-off-by: Julius Gawlas <julius.gawlas@hp.com>