client.shared.base_packages: Override checksum feature for GitFetcher
This is a solution for the issue #603.

When adding a git repo as the package provider for autotest
deps, the method BasePackageManager.install_pkg() will download
and try to verify the checksum files for the given dependency.

For example, if you use the following snippet:

job.pkgmgr.add_repository("git://10.0.0.12/tools/iometer.git")
job.setup_dep(['iometer'])

This will fail because a GitFetcher doesn't have the ability
to fetch .checksum files. So, to avoid the error, if the package
repo is a git repo, let's skip checksum verification altogether,
even overriding what the callers may have passed, since it
doesn't make sense for the git fetcher anyway.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>