server.autotest_remote: Fix autotest remote for distro package installs
As an attempt to fix problems found on our distro packages
for Fedora and RHEL, define that when autotest is running
using distro packages, it won't rsync any of the client code
to the client machines, instead, it will rely on the distro
packages themselves.

This of course reduces the usefulness of the autotest server,
since it can't test any client distro that does not have the
autotest client packaged, but this is an acceptable compromise
given the situation (frequently people using the distro
packages will be testing clients using that very same distro).

Simplify the code by ruling out rsyncs, and install/remove
autotest fom the clients using the distro supplied tools
(for now it works only for Fedora and RHEL).

Also add some more logging to the job log collector, so
we know which directories are being synced back to the
autotest server.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>