client.shared.version: Get same version name with new tagging convention
Github won't host our arbitrarily uploaded tarballs
anymore, so in order to have our upstream tarballs
to have a nicer name, we'll have to start calling
our tags differently, such as:

autotest-0.15.0

In order to tell the version name the same way as
before, strip the prepended 'autotest-'.

This commit also fixes a bug when executing the script
from an arbitrary directory outside of the git tree
dir.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>