setup.py: some adjustments to properly ship and install the right files
It was noticed that some files were not being installed / shipped
by the setuptool script. It turned out that we were using the
package data feature but providing the module directory twice.

Also, this patch removes some directories that should no longer be
shipped/installed as part of core autotest, and some that should
(installation_support).

Signed-off-by: Cleber Rosa <crosa@redhat.com>