tko: Ressurrect old TKO by re-adding the tko_test_view view
The old TKO web application (a bunch of cgi scripts located
in tko/) was broken with the change to South, since the
tko_test_view view was not contemplated in the datbase
creation schema and migration info.

Fortunately, it is all a matter of adding an sql for the
generation of the view and let autotest-database-turnkey
to process it. We don't need to add it to the Django
models or add it to the migration info, since no django
related code needs to touch it. Users that installed, say,
0.15.0 will have to manually run the sql that creates
the new view as a workaround.

CC: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>