check_patch: Compile dependency before full tree check
Make sure we've got the protobuf compiler, and then
always run a build of the protobuf module (this
module is important for some internal code inside tko,
but it is auto generated). If the package is lacking,
let the user know upfront. Point to the right docs
that will help to install all the needed packages.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>