Make it optional to install default profile
With recent changes to profile code if install server is set create job
(CLI and GUI) will default to reinstalling with current profile. This
irrited some of the users as they find it too easy to reinstall machines.
With that setting in ini file they can default to not installing
default profile.

Signed-off-by: Julius Gawlas <julius.gawlas@hp.com>