harness: new beaker harness
This is the bulk of the harness file for enabling beaker.

The idea is to bootstrap it by downloading the remote xml file, parse it,
convert it into a control file and pass that to autotest.

The rest of the harness has various hacks to deal with passing info
from autotest to beaker.

Two of the biggest pieces are converting the xml recipe into a control
file.  This will still need tweaking.

The other piece is trying to figure out when recipe/tasks/sub-tasks
start and stop.  Again probably needs tweaking.

Overall, this patch has been tested with some stub recipes and responds
well.  Work still needs to be done on figuring out which files to
upload and real workflow testing.

This patch will need to be reviewed by autotest and beaker folks.

v2: rewrote file uploading to handle reboot case (save state), python
    lint fixes, handle reservesys case, param fixes

Signed-off-by: Don Zickus <dzickus@redhat.com>