Harness: Add beaker hook for quick commands
Beaker tests call a lot of shell scripts to report statuses and logs to the beaker
server.  In order to control where the data goes and how it gets there, redirect
everything to autotest for processing.  This is done by adding a cheap hack to
the bootstrap command and creating stub scripts.

The beaker tests now call the stub scripts, which call the autotest bootstrap
command, which takes the data and reformats it for delivery to the destination.

I attached the scripts I am using as examples of usage cases.  I am trying to work
out a solution where this can be one shell script sourced from an environment.

Its a little ugly but works around some of the beaker quirks.

Signed-off-by: Don Zickus <dzickus@redhat.com>