version.py: __all__ must be a sequence of strings if defined
http://docs.python.org/2/reference/simple_stmts.html#the-import-statement

The public names defined by a module are determined by checking the
module’s namespace for a variable named __all__; if defined, it must be
a sequence of strings which are names defined or imported by that
module.

Add a , to force the single-element tuple.

This fixes an issue where epydoc was failing to parse version.py

Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>