database: increase size of TKO's test attributes value
Turns out that the definition of the test attribute value size set
in the tko models file (frontend/tko/models.py) was a lot smaller
than the size on the SQL schema file:

schema.sql snippet:

   CREATE TABLE `tko_test_attributes` (
      `value` varchar(1024) DEFAULT NULL,

frontent/tko/models.py snippet:

   class TestAttribute(dbmodels.Model, model_logic.ModelExtensions):
       value = dbmodels.CharField(blank=True, max_length=300)

This has been the case since 2009, according to commit f756779f.

So, increase the size so that attributes with larger values do not
get truncated.

Signed-off-by: Cleber Rosa <crosa@redhat.com>