parse_ssh_path: fix regex to reject urls with no netloc
parse_ssh_path("ssh:///foo") does not raise an exception.  It should
raise an exception because there is no netloc in the URL.

Replace * with + in regex, to raise an exception if there is not one or more
characters in netloc.

parse_ssh_path could be replaced with urlparse.urlsplit in Python 2.6+

Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>