RPC server: add simplest possible interface versioning
This is needed because, even if an RPC client knows the autotest
version of the RPC server, there may be differences in the interface
(think of the master and next branch).

We can not require the RPC client to parse and compare a git commit
hash, so let's introduce an interface versioning scheme.

The versioning scheme is modeled after DNS zone files version, although
for even more simplicity I decided to keep only the dates and drop
the incremental version. The obvious limitation is that this prevents
the client from differentiating two conflicting interface versions
commited upstream on the very same day (which is not at all common).

Signed-off-by: Cleber Rosa <crosa@redhat.com>