software_manager: fix check for a package and version that is not installed
When calling check_installed() and asking for a specific version
check, the current code misleadingly returns True.

The actual cause is that the output of the rpm command ("package
foo is not installed") is simply compared with the asked for version,
and the string comparison returns True.

For now just return False if that specific output is returned. A
more complete rewrite may be needed for fail proof version comparison.

Signed-off-by: Cleber Rosa <crosa@redhat.com>