tko: remove machine_load script
The fact that this script references a 'machines' database (and not
a 'tko_machines' database) shows its age. Also this specific feature
(supporting a 'machines' text file) is not documented, so I have no
idea on its intended usage.

We could fix the script, but I believe it will not be missed. If it's
missed, it'll have to be fixed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>