utils/external_packages.py: Fix Python module versions comparison.
The is_needed function was comparing Python module versions with string
comparison.

This resulted in an installation bug on Fedora 18 triggered by the
python-paramiko package. Fedora 18 is currently shipping with paramiko
1.10.1 and it compared incorrectly to the 1.7.5 version required by
Autotest.

The fix uses distutils.version.LooseVersion for Python module versions
comparison. StrictVersion cannot be used because the Autotest required
python-matplotlib version of 0.98.5.3 is an invalid version number with
StrictVersion.

Signed-off-by: Vinson Lee <vlee@twitter.com>